requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.2.3            # via django
backcall==0.1.0           # via ipython
celery==4.4.0
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
colorama==0.4.3           # via ipython
decorator==4.4.1          # via ipython, traitlets
defusedxml==0.6.0         # via python3-openid
django-allauth==0.41.0
django-autoslug==1.9.6
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django==3.0.2
djangorestframework==3.11.0
drf-yasg==1.17.0
factory-boy==2.12.0
idna==2.8                 # via requests
ipython-genutils==0.2.0   # via traitlets
ipython==7.11.1
jedi==0.15.2              # via ipython
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
parso==0.5.2              # via jedi
pickleshare==0.7.5        # via ipython
pip-tools==4.3.0
prompt-toolkit==3.0.2     # via ipython
pygments==2.5.2           # via ipython
python-dateutil==2.8.1    # via pandas
python3-openid==3.1.0     # via django-allauth
pytz==2019.3              # via django, pandas
requests-oauthlib==1.3.0  # via django-allauth
requests==2.22.0          # via django-allauth, requests-oauthlib
six==1.14.0               # via django-rest-auth, python-dateutil, traitlets
sqlparse==0.3.0           # via django
traitlets==4.3.3          # via ipython
urllib3==1.25.7           # via requests
uWSGI==2.0.18
wcwidth==0.1.8            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
